+++ app/src/main/java/com/voltek/newsfeed/data/network/entity/NewsApiArticlesResponse.kt
@@ -1,5 +1,3 @@
 package com.voltek.newsfeed.data.network.entity
 
-import com.voltek.newsfeed.data.entity.ArticleRAW
-
-data class NewsApiArticlesResponse(val articles: List<ArticleRAW>)
+data class NewsApiArticlesResponse(val articles: List<ArticleAPI>)
