+++ app/src/main/java/com/voltek/newsfeed/data/network/entity/NewsApiSourcesResponse.kt
@@ -1,5 +1,3 @@
 package com.voltek.newsfeed.data.network.entity
 
-import com.voltek.newsfeed.data.entity.SourceRAW
-
-data class NewsApiSourcesResponse(val sources: List<SourceRAW>)
+data class NewsApiSourcesResponse(val sources: List<SourceAPI>)
