+++ app/src/main/java/com/voltek/newsfeed/domain/Mapper.kt
@@ -1,14 +1,14 @@
 package com.voltek.newsfeed.domain
 
-import com.voltek.newsfeed.data.entity.ArticleRAW
-import com.voltek.newsfeed.data.entity.SourceDB
-import com.voltek.newsfeed.data.entity.SourceRAW
+import com.voltek.newsfeed.data.network.entity.ArticleAPI
+import com.voltek.newsfeed.data.storage.entity.SourceDB
+import com.voltek.newsfeed.data.network.entity.SourceAPI
 import com.voltek.newsfeed.presentation.entity.ArticleUI
 import com.voltek.newsfeed.presentation.entity.SourceUI
 
 object Mapper {
 
-    fun articleRAWtoUI(item: ArticleRAW): ArticleUI =
+    fun articleAPItoUI(item: ArticleAPI): ArticleUI =
             ArticleUI(
                     title = item.title,
                     description = item.description,
@@ -16,18 +16,7 @@ object Mapper {
                     urlToImage = item.urlToImage
             )
 
-    fun sourceRAWtoUI(item: SourceRAW): SourceUI =
-            SourceUI(
-                    id = item.id ?: "",
-                    name = item.name ?: "",
-                    description = item.description ?: "",
-                    url = item.url ?: "",
-                    category = item.category ?: "",
-                    country = item.country ?: "",
-                    isEnabled = item.isEnabled ?: false
-            )
-
-    fun sourceRAWtoDB(item: SourceRAW): SourceDB {
+    fun sourceAPItoDB(item: SourceAPI): SourceDB {
         val source = SourceDB()
         with(source) {
             id = item.id ?: ""
@@ -50,5 +39,16 @@ object Mapper {
                     category = item.category,
                     country = item.country,
                     isEnabled = item.isEnabled
+            )
+
+    fun sourceAPItoUI(item: SourceAPI): SourceUI =
+            SourceUI(
+                    id = item.id ?: "",
+                    name = item.name ?: "",
+                    description = item.description ?: "",
+                    url = item.url ?: "",
+                    category = item.category ?: "",
+                    country = item.country ?: "",
+                    isEnabled = item.isEnabled ?: false
             )
 }
