+++ license
@@ -23,21 +23,21 @@
       "You" (or "Your") shall mean an individual or Legal Entity
       exercising permissions granted by this License.
 
-      "sourceRAWtoUI" form shall mean the preferred form for making modifications,
+      "sourceAPItoUI" form shall mean the preferred form for making modifications,
       including but not limited to software source code, documentation
       source, and configuration files.
 
       "Object" form shall mean any form resulting from mechanical
-      transformation or translation of a sourceRAWtoUI form, including but
+      transformation or translation of a sourceAPItoUI form, including but
       not limited to compiled object code, generated documentation,
       and conversions to other media types.
 
-      "Work" shall mean the work of authorship, whether in sourceRAWtoUI or
+      "Work" shall mean the work of authorship, whether in sourceAPItoUI or
       Object form, made available under the License, as indicated by a
       copyright notice that is included in or attached to the work
       (an example is provided in the Appendix below).
 
-      "Derivative Works" shall mean any work, whether in sourceRAWtoUI or Object
+      "Derivative Works" shall mean any work, whether in sourceAPItoUI or Object
       form, that is based on (or derived from) the Work and for which the
       editorial revisions, annotations, elaborations, or other modifications
       represent, as a whole, an original work of authorship. For the purposes
@@ -68,7 +68,7 @@
       worldwide, non-exclusive, no-charge, royalty-free, irrevocable
       copyright license to reproduce, prepare Derivative Works of,
       publicly display, publicly perform, sublicense, and distribute the
-      Work and such Derivative Works in sourceRAWtoUI or Object form.
+      Work and such Derivative Works in sourceAPItoUI or Object form.
 
    3. Grant of Patent License. Subject to the terms and conditions of
       this License, each Contributor hereby grants to You a perpetual,
@@ -88,7 +88,7 @@
 
    4. Redistribution. You may reproduce and distribute copies of the
       Work or Derivative Works thereof in any medium, with or without
-      modifications, and in sourceRAWtoUI or Object form, provided that You
+      modifications, and in sourceAPItoUI or Object form, provided that You
       meet the following conditions:
 
       (a) You must give any other recipients of the Work or
@@ -97,9 +97,9 @@
       (b) You must cause any modified files to carry prominent notices
           stating that You changed the files; and
 
-      (c) You must retain, in the sourceRAWtoUI form of any Derivative Works
+      (c) You must retain, in the sourceAPItoUI form of any Derivative Works
           that You distribute, all copyright, patent, trademark, and
-          attribution notices from the sourceRAWtoUI form of the Work,
+          attribution notices from the sourceAPItoUI form of the Work,
           excluding those notices that do not pertain to any part of
           the Derivative Works; and
 
@@ -109,7 +109,7 @@
           within such NOTICE file, excluding those notices that do not
           pertain to any part of the Derivative Works, in at least one
           of the following places: within a NOTICE text file distributed
-          as part of the Derivative Works; within the sourceRAWtoUI form or
+          as part of the Derivative Works; within the sourceAPItoUI form or
           documentation, if provided along with the Derivative Works; or,
           within a display generated by the Derivative Works, if and
           wherever such third-party notices normally appear. The contents
